+++ pkg/handlers/resources/advanced_daemonset_handler.go
@@ -72,7 +72,7 @@ func (h *AdvancedDaemonSetHandler) Restart(c *gin.Context, namespace, name strin
 
 	result := GetKruiseOperationsManager().ExecuteOperation(c.Request.Context(), cs, req)
 	if !result.Success {
-		return errors.NewInternalError(fmt.Errorf(result.ErrorDetail))
+		return errors.NewInternalError(fmt.Errorf("%s", result.ErrorDetail))
 	}
 
 	return nil