+++ pkg/handlers/resources/cloneset_handler.go
@@ -112,7 +112,7 @@ func (h *CloneSetHandler) Restart(c *gin.Context, namespace, name string) error 
 
 	result := GetKruiseOperationsManager().ExecuteOperation(c.Request.Context(), cs, req)
 	if !result.Success {
-		return errors.NewInternalError(fmt.Errorf(result.ErrorDetail))
+		return errors.NewInternalError(fmt.Errorf("%s", result.ErrorDetail))
 	}
 
 	return nil